feat(ios): decode engine_early_stop_decision_request event
Add the engine_early_stop_decision_request RemoteEvent variant +
TypeKey case + family-decoder + encoder + DiagnosticLog case +
view-model no-op handler. iOS does not act on this event — the
desktop's early-stop-policy.ts is the authoritative responder via
the early_stop_decision_response client command — but decoding it
cleanly on iOS keeps the wire protocol uniform: every engine event
the desktop sees, iOS sees.

The case payload mirrors the Go-side EarlyStopDecisionRequestEvent
verbatim (15 associated values covering tab/instance/request IDs,
the model + turn metadata, the cumulative-output / budget /
threshold counters, the continuation count + max + last delta, the
tentative wouldContinue verdict, and the isSubagent flag). Every
wire field reaches iOS even when most are zero, so future iOS
work (e.g. a 'model nudged' status indicator) can read the full
record without contract changes.

New tests in ContractSyncTests.swift:
  - testEngineEarlyStopDecisionRequestDecode (full payload)
  - testEngineEarlyStopDecisionRequestDecodeMinimal (omitempty
    fallback verifies the decoder defaults missing values to
    zero/empty rather than throwing)

Side-effect file split: NormalizedEvent+Engine.swift hit 642/600
after the new case landed (the codec gained ~70 lines for the
new decoder + encoder + CodingKeys). Split the supporting wire
types — AgentStateUpdate, StatusFields, PermissionDenialEntry,
EngineInstancePayload, EngineMessage, EngineMessageEndUsage,
AnyCodable — into a new EngineEventSupport.swift file (~240
lines). The codec file drops to 400 lines and isolates the part
that grows with each new engine event variant. Xcode project
updated to include the new file.
