+++ ios/IonRemote/Utilities/DiagnosticLog+Events.swift
@@ -159,6 +159,16 @@ extension DiagnosticLog {
             // events.
             log("EVENT: enginePlanProposal tabId=\(tabId.prefix(8)) inst=\(instId?.prefix(8) ?? "nil") kind=\(kind) path=\(path?.suffix(40) ?? "nil")")
 
+        case .engineEarlyStopDecisionRequest(let tabId, let instId, let reqId, _, _, let turn, _, let cumOut, let budget, let pct, _, _, _, let would, _):
+            // Engine ↔ harness wire-protocol request. The desktop is the
+            // authoritative responder; iOS only observes for diagnostic
+            // visibility. Log the most useful correlation fields (request
+            // ID, turn, percent-of-budget) so a developer triaging
+            // continuation issues can pair the iOS-side log line with the
+            // engine's `earlyStop: ...` lines and the desktop's
+            // `early-stop-policy` lines.
+            log("EVENT: engineEarlyStopDecisionRequest tabId=\(tabId.prefix(8)) inst=\(instId?.prefix(8) ?? "nil") reqId=\(reqId.prefix(8)) turn=\(turn) tokens=\(cumOut)/\(budget) thr=\(pct)% would=\(would)")
+
         case .gitChangesResponse(let dir, _):
             log("EVENT: gitChangesResponse dir=\(dir.suffix(30))")
